==> 0_setup.R <==
## 0 - setup

# install required R packages 

install.packages(c("terra", "geodata", "igraph", "remotes"))

remotes::install_github("https://github.com/ElGraouiM/XSitu")
remotes::install_github("https://github.com/ElGraouiM/XSdata")
